Adjoint values of all block outputs should be reset
Not only block outputs that are dependencies of blocks on the tape, but
also those that are outputs need to be reset. This is because these
outputs may serve as a functional, in which case its adjoint value will
be set to 1.0 at the beginning of the adjoint calculation. If
subsequently the gradient for a *different* functional is evaluated, not
only the adj_value of that new value will start at 1.0, but adj_value of
the previous functional is also still 1.0, and so the gradient
calculation will percolate the sensitivity of  both outputs backward.