Fix vectorial adjoint values for firedrake_adjoint
Vectorial adjoint values (cofunctions) are now stored as Vector, just
like in fenics_adjoint, instead of Functions. This is to ensure when these
values are added they are added vectorially, instead of ending up with
UFL expressions of linear combinations of Functions.

To implement this I have added a new compat.assemble_adjoint_value()
which is a thin wrapper around backend.assemble() that ensures a Vector is
returned when assembling a 1-form. This should be called when assembling
any tlm, adjoint or hessian value that may result in a vector.

Fixes issue #28