dirichletbc: change DirichletBC checkpoints to be weak
This means that checkpoints of DirichletBC is not a new instance,
but just enough data to set the DirichletBC back to the state it was in.
Thus we only need to store the DirichletBC value BlockOutput and then
set the value of DirichletBC to the value BlockOutput.checkpoint.

This solves a problem with Expressions in DirichletBC not being
properly annotated after optimizing tape.