function: dirichletbc: fix seemingly random parallel bug
Seems I should have respected the warning about only internal library use.
By the looks of it, doing Function(V, vector) in parallel has undefined behaviour.
As we only invoked this for the adjoint solution on the boundary, it only affected dirichletbc controls in parallel.
This resulted in adjoint DirichletBC.apply (sometimes) giving different values between each apply.

I do not see the same warning in the firedrake Function docstring. Thus I will leave it as it is.