overloaded type: remove original block variable reference
There is no reason to store a reference to the first block variable, it should not be considered special and might even be unrelated to the current tape or problem.
User access to block variables should ideally be through the Control interface as this uniquely determines the block variable of interest.

See issue #25.