Several improvement to Function.assign
Corrects adjoint of assignment of Constant to non-Lagrangian functions.
Ensures that Constants in assignment to a Function, produce a
Vector-valued adj_value associated with the R-space associated with that
constant.
Reversely, in the adjoint of an assignment of an AdjFloat to a Constant,
the Vector-valued adj_value is converted back to a float.

Passes all firedrake tests.