Teach ROL Objective about update flags
ROL 2.0 calls the objective update method with a flag describing why
the call was taken. This allows us to implement a caching scheme,
to accept or revert updates if necessary. Ultimately, this can save
quite a few objective function evaluations.

Because the ROL 2.0 update is very widely split, this change is
backwards compatible, and will only be effective if pyrol wraps the
UpdateType enumeration.