Fix error message for post-derivative calbback
It was copied from the pre-derivative callback in the wrong place, and
without adapting the message.

I also think they should be errors instead
of just DeprecationWarnings for both the pre and post callback: if the
controls get set to None by the callback, it will definitely crash later
on anyway, and if the post callback sets the derivatives to None, the
subsequent delist will make it return the control values instead of the
derivative - which is definitely what you'd ever want but is hard to
debug as you get something of the right shape, just with the wrong
values.